LayoutTests/imported/w3c/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/navigatorcookies-cookieenabled-true-expected.txt


PASS navigator.cookieEnabled is true when cookies are enabled
